better format for component documentation
(I just updated the firs part to give the example)

I like the idea of the Readme.md documentation, but I am not a big fan of the table-oriented format you used, because it makes the source text of the Readme hard to read and write. Readable source-code is one the big pluses of markdown.

In addition, the table-oriented format is not git-friendly, because it forces you to cram much info on a single line.

The line-oriented format I proposed is better in those regards. Plus, by using subtitles for each attribute and event, you get a direct link to each of them for free.

I can update the rest of the file if you agree with this change.

The list-oriented format I proposed is better in that re
The format (using subtitles for each attribute and event) I proposed here has several advantages:

* easier to read in the source code and to edit
* more git-friendly (because the content is